requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
click==8.1.2
dapr==1.6.0
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.1
frozenlist==1.3.0
greenlet==1.1.2
grpcio==1.44.0
h11==0.13.0
idna==3.3
motor==2.5.1
multidict==6.0.2
protobuf==3.19.4
py-bcrypt==0.4
pydantic==1.9.0
pymongo==3.12.3
python-dateutil==2.8.2
python-decouple==3.6
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
yarl==1.7.2
